Only retry for server auth error
