+++ src/commands/storageAccountActionHandlers.ts
@@ -3,7 +3,6 @@
  *  Licensed under the MIT License. See License.md in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as retry from 'p-retry';
 import * as vscode from "vscode";
 import { IActionContext, registerCommand } from 'vscode-azureextensionui';
 import { configurationSettingsKeys, extensionPrefix } from '../constants';
@@ -13,7 +12,6 @@ import { BlobContainerTreeItem } from "../tree/blob/BlobContainerTreeItem";
 import { StorageAccountTreeItem } from '../tree/StorageAccountTreeItem';
 import { isPathEqual, isSubpath } from '../utils/fs';
 import { localize } from "../utils/localize";
-import { nonNullValue } from '../utils/nonNull';
 import { showWorkspaceFoldersQuickPick } from "../utils/quickPickUtils";
 import { deleteNode } from './commonTreeCommands';
 import { selectStorageAccountTreeItemForCommand } from './selectStorageAccountNodeForCommand';
@@ -96,18 +94,7 @@ async function deployStaticWebsite(context: IActionContext, target?: vscode.Uri 
 
     await runPreDeployTask(sourcePath, context);
 
-    const retries: number = 4;
-    await retry(
-        async (currentAttempt) => {
-            context.telemetry.properties.deployAttempt = currentAttempt.toString();
-            if (currentAttempt > 1) {
-                const message: string = localize('retryingDeploy', 'Retrying deploy (Attempt {0}/{1})...', currentAttempt, retries + 1);
-                ext.outputChannel.appendLog(message);
-            }
-            await nonNullValue(destContainerTreeItem).deployStaticWebsite(context, nonNullValue(sourcePath));
-        },
-        { retries, minTimeout: 2 * 1000 }
-    );
+    return destContainerTreeItem.deployStaticWebsite(context, sourcePath);
 }
 
 async function runPreDeployTask(deployFsPath: string, context: IActionContext): Promise<void> {
